agora só quem é admin consegue postar/editar e deletar

+++ main.py
@@ -18,22 +18,20 @@ class User(UserMixin, db.Model):
     username = db.Column(db.String(50), unique=True, nullable=False)
     password_hash = db.Column(db.String(100), nullable=False)
     admin = db.Column(db.Boolean, default=False)
+    
 
     def set_password(self, password):
         self.password_hash = generate_password_hash(password)
 
     def check_password(self, password):
         return check_password_hash(self.password_hash, password)
-    
-  
-
-   
 
 # Modelo de post para o banco de dados
 class Post(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(100), nullable=False)
     content = db.Column(db.Text, nullable=False)
+  
 
 # Crie o banco de dados e as tabelas dentro do contexto do aplicativo
 with app.app_context():
@@ -47,16 +45,19 @@ def load_user(user_id):
 # Rota principal - lista de posts
 @app.route('/')
 def index():
-    nome = request.args.get("nome")
-    posts = Post.query.all()
-    return render_template("index.html", nome=nome, posts=posts)
+    if current_user.is_authenticated:
+        nome = request.args.get("nome")
+        posts = Post.query.all()
+        return render_template("index.html", nome=nome, posts=posts)
+    else:
+        return redirect(url_for('login'))
 
 # Rota para adicionar um novo post
 @app.route('/add_post', methods=['POST'])
 @login_required
 def add_post():
     data = request.get_json()
-    new_post = Post(title=data['title'], content=data['content'])
+    new_post = Post(title=data['title'], content=data['content'], user_id=current_user.id)
     db.session.add(new_post)
     db.session.commit()
     return jsonify({"message": "Post added"}), 201
@@ -67,6 +68,9 @@ def add_post():
 def delete_post(id):
     post = Post.query.get(id)
     if post:
+        if post.author != current_user and not current_user.admin:
+            return jsonify({"message": "Unauthorized"}), 403
+        
         db.session.delete(post)
         db.session.commit()
         return jsonify({"message": "Post deleted"}), 200
@@ -79,6 +83,9 @@ def edit_post(id):
     data = request.get_json()
     post = Post.query.get(id)
     if post:
+        if post.author != current_user and not current_user.admin:
+            return jsonify({"message": "Unauthorized"}), 403
+        
         post.title = data['title']
         post.content = data['content']
         db.session.commit()
